refactor: 가독성을 위한 패키지 분리
