+++ backend/bottari/src/main/java/com/bottari/push/notification/reliablefcm/domain/FcmSendTask.java
@@ -1,4 +1,4 @@
-package com.bottari.push.notification.reliablefcm;
+package com.bottari.push.notification.reliablefcm.domain;
 
 import com.bottari.error.BusinessException;
 import com.bottari.error.ErrorCode;
