+++ backend/bottari/src/main/java/com/bottari/push/notification/reliablefcm/repository/FcmSendTaskRepository.java
@@ -1,5 +1,6 @@
-package com.bottari.push.notification.reliablefcm;
+package com.bottari.push.notification.reliablefcm.repository;
 
+import com.bottari.push.notification.reliablefcm.domain.FcmSendTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FcmSendTaskRepository extends JpaRepository<FcmSendTask, Long> {
