+++ backend/bottari/src/main/java/com/bottari/push/notification/reliablefcm/repository/FcmSendTaskStateRepository.java
@@ -1,5 +1,6 @@
-package com.bottari.push.notification.reliablefcm;
+package com.bottari.push.notification.reliablefcm.repository;
 
+import com.bottari.push.notification.reliablefcm.domain.FcmSendTaskState;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FcmSendTaskStateRepository extends JpaRepository<FcmSendTaskState, Long> {
