+++ backend/bottari/src/test/java/com/bottari/push/notification/reliablefcm/FcmSendTaskServiceTest.java
@@ -4,6 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.bottari.push.message.PushMessage;
+import com.bottari.push.notification.reliablefcm.domain.FcmSendTaskState;
+import com.bottari.push.notification.reliablefcm.domain.TaskState;
+import com.bottari.push.notification.reliablefcm.dto.ScheduleFcmSendTaskRequest;
+import com.bottari.push.notification.reliablefcm.service.FcmSendTaskService;
 import jakarta.persistence.EntityManager;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
